# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vello Trade API
  description: |
    REST API for the Vello Trade B2B marketplace mobile app (Expo). JSON routes are mounted under **`/api`** (the mobile client uses `baseURL: {host}/api`).

    **Socket.IO** runs on the same HTTP server. After login, connect with JWT in handshake `auth: { token: "<jwt>" }`.
    Events include `room:join`, `chat:typing`, `message:new`, `chat:room:updated`. See the repository README for details.
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/password:
    patch:
      tags: [Auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
...
  /api/auth/profile:
    get:
      tags: [Auth]
      summary: Current user (alias)
    patch:
      tags: [Auth]
      summary: Update display name only
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
...